www/requirements.txt
django
psycopg2
#postgis
djangorestframework
pyyaml
django-filter
coreapi
markdown
django-crispy-forms
django-guardian
